+++ app/__main__.py
@@ -1,65 +1,37 @@
 #!/usr/bin/env python
-"""Redirects to entrypoint of the app"""
+"""
+Sets state of the app and redirects to the entrypoint.
 
+The state of the app is comprised of
+- App mode, i.e. ["train"|"infer"]
+- Global state of logging
+- Showing system information
+"""
 
-def _parse_args() -> tuple[str, bool, bool]:
-    """
-    Parses argv and returns the inputs.
+from sys import exit
 
-    Returns tuple of
-    - String of app mode out of `["train", "infer"]`
-    - Tuple of bools to output debug and/or sysinfo
-    """
+from .helper.get_and_configure_logger import (
+    get_and_configure_logger,
+    toggle_global_debug_state,
+)
+from .helper.get_and_configure_system_info import toggle_global_sysinfo
+from .helper.parse_args import parse_args
 
-    desc = "Create pipeline object parametrised with parameter object and execute task."
-    modes = ["train", "infer"]
+if __name__ == "__main__":
 
-    parser = ArgumentParser(prog="app", description=desc)
-    parser.add_argument(
-        "-m",
-        "--mode",
-        type=str,
-        default=modes[0],
-        choices=modes,
-        help="change app mode",
-    )
-    parser.add_argument(
-        "-d", "--debug", action="store_true", help="show debug messages"
-    )
-    parser.add_argument(
-        "-i", "--sysinfo", action="store_true", help="show system information"
-    )
-    args = parser.parse_args()
-
-    return (args.mode, args.debug, args.sysinfo)
-
-
-def _set_global_state(show_debug: bool = False, show_sysinfo: bool = False):
-    """Sets the global state of logging and showing system information"""
+    app_mode, show_debug, show_sysinfo = parse_args()
 
     try:
         toggle_global_debug_state(show_debug)
         toggle_global_sysinfo(show_sysinfo)
     except Exception as e:
-        get_and_configure_logger(__name__).error(e)
-        return e
+        exit(get_and_configure_logger(__name__).error(e))
 
-
-if __name__ == "__main__":
-
-    from argparse import ArgumentParser
-    from sys import exit
-
-    from .helper.get_and_configure_logger import (
-        get_and_configure_logger,
-        toggle_global_debug_state,
-    )
-    from .helper.get_and_configure_system_info import toggle_global_sysinfo
-
-    app_mode, show_debug, show_sysinfo = _parse_args()
-    _set_global_state(show_debug, show_sysinfo)
-
-    # delayed import to account for logger set to be first
+    # TODO delayed import to account for logger set to be first
     from .app import main
 
     exit(main(app_mode))
+
+else:
+    msg = "Not inside __main__. Exiting."
+    exit(get_and_configure_logger(__name__).error(msg))